Notifications

+++ app/Jobs/UpdateInstallPlugins.php
@@ -13,6 +13,14 @@ use GitWrapper\{GitWrapper, GitWorkingCopy};
 use GitWrapper\Exception\GitException;
 
 use App\Helpers\Proc;
+use App\Helpers\Notify;
+use App\Notifications\PluginUpdateError;
+use App\Notifications\InstallCloneError;
+use App\Notifications\InstallPrError;
+use App\Notifications\InstallDeployError;
+use App\Notifications\InstallTagError;
+use App\Notifications\InstallUpdateReport;
+use App\Notifications\InstallNoUpdates;
 use App\Install;
 
 class UpdateInstallPlugins implements ShouldQueue {
@@ -22,14 +30,16 @@ class UpdateInstallPlugins implements ShouldQueue {
 	const GIT_EMAIL = 'dev@viastudio.com';
 
 	protected $installId;
+	protected $opts;
 
 	/**
 	 * Create a new job instance.
 	 *
 	 * @return void
 	 */
-	public function __construct(int $installId) {
+	public function __construct(int $installId, array $opts = []) {
 		$this->installId = $installId;
+		$this->opts = $opts;
 	}
 
 	/**
@@ -43,7 +53,7 @@ class UpdateInstallPlugins implements ShouldQueue {
 
 		$git = new GitWrapper();
 		$git->setTimeout(600); //10 minutes
-		//TODO - disable this when we're done?
+		//TODO - disable this when we're done or make it an option?
 		$git->streamOutput();
 
 		$baseUrl = 'git@github.com:viastudio/%s.git';
@@ -59,7 +69,6 @@ class UpdateInstallPlugins implements ShouldQueue {
 			try {
 				$tries++;
 				if ($tries > $max) {
-					//TODO - exception
 					Log::error(
 						"{$install->name} exceeded number of clone retries",
 					);
@@ -81,7 +90,6 @@ class UpdateInstallPlugins implements ShouldQueue {
 					);
 					$repoUrl = sprintf($baseUrl, $domain);
 				} else {
-					//TODO - maybe notification?
 					Log::error(
 						"Error cloning {$install->name}: " . $e->getMessage(),
 					);
@@ -89,13 +97,13 @@ class UpdateInstallPlugins implements ShouldQueue {
 			}
 		}
 
-		//We exceeded the number of tries to clone so we move on to the next site
+		//We exceeded the number of tries to clone so we abort updating this site
 		if ($tries > $max) {
-			//TODO - exception/notification;
+			Notify::send(new InstallCloneError($install));
 			return;
 		}
 
-		//TODO start updating plugins
+		// Start updating plugins
 		$today = new \DateTime();
 		$tstamp = $today->format('Y-m-d-H-i-s');
 		$branch = "auto-plugin-updates-{$tstamp}";
@@ -168,11 +176,22 @@ class UpdateInstallPlugins implements ShouldQueue {
 
 				$pluginsUpdated[] = $plugin;
 			} catch (\Exception $e) {
-				Log::error($e->getMessage());
+				Notify::send(new PluginUpdateError($e->getMessage()));
+				return;
 			}
 		}
 
+		Log::debug('Finished updating plugins');
+		Log::debug($pluginsUpdated);
+		Log::debug($pluginsSkipped);
+
+		if (count($pluginsUpdated) <= 0) {
+			Notify::send(new InstallNoUpdates($install));
+			return;
+		}
+
 		try {
+			$prCreated = false;
 			// PR and merge
 			$process = Proc::exec(
 				[
@@ -189,36 +208,49 @@ class UpdateInstallPlugins implements ShouldQueue {
 
 			sleep(30); // Give the PR time to finish setting up
 			$process = Proc::exec(['gh', 'pr', 'merge', '--merge'], $wcPath);
+			$prCreated = true;
+			Log::debug('Finished creating PR');
 		} catch (\Exception $e) {
+			$msg = $e->getMessage();
+
 			Log::error('PR error: ' . $e->getMessage());
+			Notify::send(new InstallPrError($install, $e->getMessage()));
 		}
 
-		try {
-			//Push to WPE
-			$wc->remote(
-				'add',
-				'production',
-				"git@git.wpengine.com:production/{$install->name}.git",
-			);
-			Proc::exec(['git', 'clean', '-d', '-fx'], $wcPath);
-			Proc::exec(
-				[
-					'git',
-					'pull',
-					'-X',
-					'theirs',
-					'--no-edit',
+		if ($prCreated) {
+			try {
+				//Push to WPE
+				$wc->remote(
+					'add',
 					'production',
-					'master',
-				],
-				$wcPath,
-			);
-			$wc->push('production', 'master', [
-				'force' => true,
-				'no-verify' => true,
-			]);
-		} catch (\Exception $e) {
-			Log::error('Production push error: ' . $e->getMessage());
+					"git@git.wpengine.com:production/{$install->name}.git",
+				);
+				Proc::exec(['git', 'clean', '-d', '-fx'], $wcPath);
+				Proc::exec(
+					[
+						'git',
+						'pull',
+						'-X',
+						'theirs',
+						'--no-edit',
+						'production',
+						'master',
+					],
+					$wcPath,
+				);
+				$wc->push('production', 'master', [
+					'force' => true,
+					'no-verify' => true,
+				]);
+				Log::debug('Finished deploying to production');
+			} catch (\Exception $e) {
+				Log::error('Production push error: ' . $e->getMessage());
+				Notify::send(
+					new InstallDeployError($install, $e->getMessage()),
+				);
+			}
+		} else {
+			Log::debug('Skipping production deploy because there was no PR');
 		}
 
 		try {
@@ -227,6 +259,11 @@ class UpdateInstallPlugins implements ShouldQueue {
 			$wc->pushTag($tagName);
 		} catch (\Exception $e) {
 			Log::error('Tag error: ' . $e->getMessage());
+			Notify::send(new InstallTagError($install, $e->getMessage()));
 		}
+
+		Notify::send(
+			new InstallUpdateReport($install, $pluginsUpdated, $pluginsSkipped),
+		);
 	}
 }
